Create a file for helper classes
